Adding primitive support for custom javascript bundlers
I took a look and think this is a basic solution for extending the existing Bundle() method to allow some custom bundlers.  I'm not happy about having to know about the internal workings of the JavaScriptBundle Render method in order to stop the hashkey from appending, but I think it's a good stopgap solution that doesn't require you to change much other code.

I added some generic methods for creating the JavaScriptBundle from the Factory and refactored the existing Create() to use them.  I made sure the unit tests still passed and added a couple that should give a basic testing of the new custom bundle functionality.  There weren't any BundleTests so I didn't bother to create them, but I did add an example in the SquishItAspNetTest Global.asax.cs below the existing Bundle() (I commented it out so there wouldn't be any surprises).