Added test for checking multiple less compilation on seperate threads.
This is to try and trigger a bug where the processing different less files
in different directories at the same time can cause them to pick up the
wrong import files because the current directory is changed on a per
process basis when preprocessing the files.

This is a race issue, so may not occur every time, but on a multicore
machine this test seems to have a high failure rate.