Added "AutoBundling" feature code (did not enable)
I've dropped off code for the AutoBundle feature we discussed. I
recently uninstalled VS2010, so I did not commit my changes to the
project file (1 assembly reference and 2 new files in project).
Additionally I commented the methods added to
SquishIt.Mvc.HtmlHelperExtensions. Also note the code is currently
inaccurate in the order in which it emits resource links. I'm in the
process of modifying how strings are collated to support removing
duplicate bundles.